refactor(services): extract web prompt hook client
Fourteen prompt hook methods and the two `let` bindings behind them
(`memoryPromptHooks`, `memoryPromptTraces`) move out of the composition root.
The bindings stay unexported: `web-prompt-hook-store.ts` owns them and exposes
behavior, so there is no mutable binding two modules could re-import into
divergent copies of the mock world.

The store splits from the seeds because both together exceed the 300-line rule.

Also lands the two shared helpers this split needs. `web-mock-clock.ts` holds
`nowIso`/`daysAgoIso`, which 53 call sites depend on and which must stay one
function rather than being copied per module. `web-mock-storage.ts` is the single
storage entry point task 3.1 calls for; the prompt hook pair is its first caller
and preserves the same keys, the same in-memory fallback on a missing or
unparseable entry, and the same 50-trace write cap.

The `src/services` aggregate rises 18,149 -> 18,195. That is the fixed per-module
cost of eight new files — imports, the `export const` line, the closing brace.
Method bodies are moved verbatim and the narrow interfaces' signatures move out
of `AgentService` rather than being duplicated, so nothing here is a rewrite.
